decompose and package project
